+++ flink-optimizer/src/test/java/org/apache/flink/optimizer/plantranslate/JobGraphGeneratorTest.java
@@ -41,6 +41,7 @@ import org.apache.flink.runtime.io.network.partition.ResultPartitionType;
 import org.apache.flink.runtime.jobgraph.IntermediateDataSetID;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobgraph.JobVertex;
+import org.apache.flink.runtime.jobgraph.generator.JobGraphUtils;
 import org.apache.flink.util.AbstractID;
 
 import org.hamcrest.Matchers;
@@ -245,7 +246,7 @@ public class JobGraphGeneratorTest {
 			Tuple2.of(nonExecutableDirName, new DistributedCache.DistributedCacheEntry(directory2.toString(), false))
 		);
 
-		JobGraphGenerator.addUserArtifactEntries(originalArtifacts, jb);
+		JobGraphUtils.addUserArtifactEntries(originalArtifacts, jb);
 
 		Map<String, DistributedCache.DistributedCacheEntry> submittedArtifacts = jb.getUserArtifacts();
